examples/components/pyproject.toml
[project]
name = "components"
requires-python = ">=3.11.10"
version = "0.1.0"
dependencies = [
    "dagster",
    "dagster-datacontract"
]

[project.entry-points]
"dagster_dg.plugin" = { components = "components.lib"}

[dependency-groups]
dev = [
    "dagster-webserver",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "components"
python_environment = "persistent_uv"
